requirements.txt
streamlit
streamlit-folium
folium
google-generativeai
httpx
Pillow
plotly
